Refactor: extract stickiness option calculation into separate hook
